// frontend-app/main-page/src/model/FilteringParamsEnum.ts
enum FilteringParamsEnum {
    BRAND_FIELD_FILTER = 'brand',
    MODEL_FIELD_FILTER = 'model',
    BODY_TYPE_FIELD_FILTER = 'bodyType',
    COLOR_FIELD_FILTER = 'color',
    DAILY_FEE_FROM_FILTER = 'dailyFeeFrom',
    DAILY_FEE_TO_FILTER = 'dailyFeeTo',
    SEATS_NUMBER_FROM_FILTER = 'seatsNumberFrom',
    DOORS_NUMBER_FROM_FILTER = 'doorsNumberFrom',
    PRODUCTION_YEAR_FROM_FILTER = 'productionYearFrom',
    SEATS_NUMBER_TO_FILTER = 'seatsNumberTo',
    DOORS_NUMBER_TO_FILTER = 'doorsNumberTo',
    PRODUCTION_YEAR_TO_FILTER = 'productionYearTo',
}

export default FilteringParamsEnum;